+++ home-health-care/src/main/java/it/univpm/hhc/services/SubService.java
@@ -10,7 +10,7 @@ public interface SubService {
 	
 	Sub findById(Long id);
 	
-	Sub create( String name, Double price, int discount);
+	Sub create( String name, Long price, int discount);
 	
 	Sub update(Sub sub);
